refactor: restructure db and cache pkg
