+++ pkg/cache/chat/get_history_test.go
@@ -1,0 +1,67 @@
+package chatcache
+
+import (
+	"context"
+	"encoding/json"
+	"testing"
+
+	"github.com/go-redis/redismock/v9"
+	"github.com/stretchr/testify/assert"
+
+	"github.com/ACaiCat/tiktok-go/pkg/db/model"
+)
+
+func TestChatCache_GetChatHistory(t *testing.T) {
+	type testCase struct {
+		userID      int64
+		otherUserID int64
+		pageSize    int
+		pageNum     int
+		stored      []*model.ChatMessage
+		mockErr     error
+		wantErr     bool
+	}
+
+	testCases := map[string]testCase{
+		"get history success": {
+			userID:      1,
+			otherUserID: 2,
+			pageSize:    10,
+			pageNum:     0,
+			stored:      testMessages,
+			wantErr:     false,
+		},
+		"cache miss returns error": {
+			userID:      1,
+			otherUserID: 2,
+			pageSize:    10,
+			pageNum:     0,
+			mockErr:     assert.AnError,
+			wantErr:     true,
+		},
+	}
+
+	for name, tc := range testCases {
+		t.Run(name, func(t *testing.T) {
+			db, mock := redismock.NewClientMock()
+			cache := NewChatCache(db)
+
+			key := getHistoryKey(tc.userID, tc.otherUserID, tc.pageSize, tc.pageNum)
+			if tc.mockErr != nil {
+				mock.ExpectGet(key).SetErr(tc.mockErr)
+			} else {
+				data, _ := json.Marshal(tc.stored)
+				mock.ExpectGet(key).SetVal(string(data))
+			}
+
+			msgs, err := cache.GetChatHistory(context.Background(), tc.userID, tc.otherUserID, tc.pageSize, tc.pageNum)
+			if tc.wantErr {
+				assert.Error(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.Len(t, msgs, len(tc.stored))
+			assert.NoError(t, mock.ExpectationsWereMet())
+		})
+	}
+}
