+++ pkg/cache/chat/set_history_test.go
@@ -2,11 +2,78 @@ package chatcache
 
 import (
 	"context"
+	"encoding/json"
 	"testing"
 
 	"github.com/go-redis/redismock/v9"
 	"github.com/stretchr/testify/assert"
+
+	"github.com/ACaiCat/tiktok-go/pkg/constants"
+	"github.com/ACaiCat/tiktok-go/pkg/db/model"
 )
+
+func TestChatCache_SetChatHistory(t *testing.T) {
+	type testCase struct {
+		userID      int64
+		otherUserID int64
+		pageSize    int
+		pageNum     int
+		messages    []*model.ChatMessage
+		mockErr     error
+		wantErr     bool
+	}
+
+	testCases := map[string]testCase{
+		"set history success": {
+			userID:      1,
+			otherUserID: 2,
+			pageSize:    10,
+			pageNum:     0,
+			messages:    testMessages,
+			wantErr:     false,
+		},
+		"redis error returns error": {
+			userID:      1,
+			otherUserID: 2,
+			pageSize:    10,
+			pageNum:     0,
+			messages:    testMessages,
+			mockErr:     assert.AnError,
+			wantErr:     true,
+		},
+		"key is normalized by user order": {
+			userID:      2,
+			otherUserID: 1,
+			pageSize:    10,
+			pageNum:     0,
+			messages:    testMessages,
+			wantErr:     false,
+		},
+	}
+
+	for name, tc := range testCases {
+		t.Run(name, func(t *testing.T) {
+			db, mock := redismock.NewClientMock()
+			cache := NewChatCache(db)
+
+			key := getHistoryKey(tc.userID, tc.otherUserID, tc.pageSize, tc.pageNum)
+			data, _ := json.Marshal(tc.messages)
+			if tc.mockErr != nil {
+				mock.ExpectSet(key, data, constants.ChatHistoryCacheExpiration).SetErr(tc.mockErr)
+			} else {
+				mock.ExpectSet(key, data, constants.ChatHistoryCacheExpiration).SetVal("OK")
+			}
+
+			err := cache.SetChatHistory(context.Background(), tc.userID, tc.otherUserID, tc.pageSize, tc.pageNum, tc.messages)
+			if tc.wantErr {
+				assert.Error(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.NoError(t, mock.ExpectationsWereMet())
+		})
+	}
+}
 
 func TestChatCache_ClearChatHistory(t *testing.T) {
 	type testCase struct {
