+++ pkg/cache/user/get_like_test.go
@@ -1,0 +1,117 @@
+package usercache
+
+import (
+	"context"
+	"testing"
+
+	"github.com/go-redis/redismock/v9"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestUserCache_GetLikedVideos(t *testing.T) {
+	type testCase struct {
+		userID  int64
+		stored  []string
+		mockErr error
+		wantLen int
+		wantErr bool
+	}
+
+	testCases := map[string]testCase{
+		"get liked videos success": {
+			userID:  1,
+			stored:  []string{"10", "20", "30"},
+			wantLen: 3,
+			wantErr: false,
+		},
+		"redis error returns error": {
+			userID:  1,
+			mockErr: assert.AnError,
+			wantErr: true,
+		},
+		"empty set returns empty slice": {
+			userID:  1,
+			stored:  []string{},
+			wantLen: 0,
+			wantErr: false,
+		},
+	}
+
+	for name, tc := range testCases {
+		t.Run(name, func(t *testing.T) {
+			db, mock := redismock.NewClientMock()
+			cache := NewUserCache(db)
+
+			key := getLikedVideosKey(tc.userID)
+			if tc.mockErr != nil {
+				mock.ExpectSMembers(key).SetErr(tc.mockErr)
+			} else {
+				mock.ExpectSMembers(key).SetVal(tc.stored)
+			}
+
+			ids, err := cache.GetLikedVideos(context.Background(), tc.userID)
+			if tc.wantErr {
+				assert.Error(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.Len(t, ids, tc.wantLen)
+			assert.NoError(t, mock.ExpectationsWereMet())
+		})
+	}
+}
+
+func TestUserCache_IsVideoLiked(t *testing.T) {
+	type testCase struct {
+		userID  int64
+		videoID int64
+		result  bool
+		mockErr error
+		wantErr bool
+	}
+
+	testCases := map[string]testCase{
+		"video is liked": {
+			userID:  1,
+			videoID: 10,
+			result:  true,
+			wantErr: false,
+		},
+		"video is not liked": {
+			userID:  1,
+			videoID: 99,
+			result:  false,
+			wantErr: false,
+		},
+		"redis error returns error": {
+			userID:  1,
+			videoID: 10,
+			mockErr: assert.AnError,
+			wantErr: true,
+		},
+	}
+
+	for name, tc := range testCases {
+		t.Run(name, func(t *testing.T) {
+			db, mock := redismock.NewClientMock()
+			cache := NewUserCache(db)
+
+			key := getLikedVideosKey(tc.userID)
+			mock.ExpectExists(getFollowingKey(tc.userID)).SetVal(1)
+			if tc.mockErr != nil {
+				mock.ExpectSIsMember(key, tc.videoID).SetErr(tc.mockErr)
+			} else {
+				mock.ExpectSIsMember(key, tc.videoID).SetVal(tc.result)
+			}
+
+			liked, err := cache.IsVideoLiked(context.Background(), tc.userID, tc.videoID)
+			if tc.wantErr {
+				assert.Error(t, err)
+				return
+			}
+			assert.NoError(t, err)
+			assert.Equal(t, tc.result, liked)
+			assert.NoError(t, mock.ExpectationsWereMet())
+		})
+	}
+}
