+++ pkg/cache/user/set_jwch_test.go
@@ -62,56 +62,6 @@ func TestUserCache_SetJwchSession(t *testing.T) {
 	}
 }
 
-func TestUserCache_GetJwchSession(t *testing.T) {
-	type testCase struct {
-		userID  int64
-		jwchID  string
-		cookie  string
-		mockErr error
-		wantErr bool
-	}
-
-	testCases := map[string]testCase{
-		"get jwch session success": {
-			userID:  1,
-			jwchID:  "jwch-id-001",
-			cookie:  "session=abc",
-			wantErr: false,
-		},
-		"key not found returns error": {
-			userID:  1,
-			mockErr: assert.AnError,
-			wantErr: true,
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			db, mock := redismock.NewClientMock()
-			cache := NewUserCache(db)
-
-			key := getJwchSessionKey(tc.userID)
-			if tc.mockErr != nil {
-				mock.ExpectGet(key).SetErr(tc.mockErr)
-			} else {
-				session := &jwchSession{ID: tc.jwchID, Cookie: tc.cookie}
-				data, _ := json.Marshal(session)
-				mock.ExpectGet(key).SetVal(string(data))
-			}
-
-			id, cookie, err := cache.GetJwchSession(context.Background(), tc.userID)
-			if tc.wantErr {
-				assert.Error(t, err)
-				return
-			}
-			assert.NoError(t, err)
-			assert.Equal(t, tc.jwchID, id)
-			assert.Equal(t, tc.cookie, cookie)
-			assert.NoError(t, mock.ExpectationsWereMet())
-		})
-	}
-}
-
 func TestUserCache_CleanJwchSession(t *testing.T) {
 	type testCase struct {
 		userID  int64
