+++ pkg/cache/user/user.go
@@ -8,6 +8,11 @@ type UserCache struct {
 	c *redis.Client
 }
 
+type jwchSession struct {
+	ID     string
+	Cookie string
+}
+
 func NewUserCache(cache *redis.Client) *UserCache {
 	return &UserCache{c: cache}
 }
