+++ pkg/constants/video.go
@@ -7,6 +7,7 @@ const (
 	DefaultVideoPageSize        = 10
 	MaxVideoPageSize            = 50
 	VideoCacheExpiration        = 24 * time.Hour
+	UserVideoCacheExpiration    = 5 * time.Minute
 	PopularVideoCacheExpiration = time.Minute * 1
 	PopularVideoCacheCount      = 1000
 )
