+++ pkg/db/comment/get_comment.go
@@ -60,3 +60,18 @@ func (c *CommentDao) GetCommentsByCommentID(ctx context.Context, commentID int64
 
 	return comments, nil
 }
+
+func (c *CommentDao) IsCommentExists(ctx context.Context, commentID int64) (bool, error) {
+	var err error
+
+	count, err := c.q.Comment.WithContext(ctx).
+		Select(c.q.Comment.ID).
+		Where(c.q.Comment.ID.Eq(commentID)).
+		Count()
+
+	if err != nil {
+		return false, errors.Wrapf(err, "IsCommentExists failed, commentID: %d", commentID)
+	}
+
+	return count > 0, nil
+}
