+++ pkg/db/comment/get_comment_test.go
@@ -121,3 +121,41 @@ func TestCommentDao_GetCommentsByCommentID(t *testing.T) {
 		})
 	}
 }
+
+func TestCommentDao_IsCommentExists(t *testing.T) {
+	type testCase struct {
+		commentID int64
+		mockRet   bool
+		mockErr   error
+		wantErr   bool
+	}
+
+	testCases := map[string]testCase{
+		"comment exists":         {commentID: 1, mockRet: true},
+		"comment not exists":     {commentID: 99, mockRet: false},
+		"db error returns error": {commentID: 1, mockErr: assert.AnError, wantErr: true},
+	}
+
+	defer mockey.UnPatchAll()
+
+	for name, tc := range testCases {
+		mockey.PatchConvey(name, t, func() {
+			mockCommentQueryChain()
+			dao := newTestDao()
+			count := int64(0)
+			if tc.mockRet {
+				count = 1
+			}
+			dbtestutil.MockCount(count, tc.mockErr)
+
+			ok, err := dao.IsCommentExists(context.Background(), tc.commentID)
+			if tc.wantErr {
+				assert.Error(t, err)
+				assert.ErrorContains(t, err, "IsCommentExists failed")
+			} else {
+				assert.NoError(t, err)
+				assert.Equal(t, tc.mockRet, ok)
+			}
+		})
+	}
+}
