+++ pkg/db/follower/get_follow.go
@@ -9,6 +9,8 @@ import (
 	"github.com/ACaiCat/tiktok-go/pkg/db/query"
 )
 
+const mutualFollowCount = 2
+
 func (f *FollowerDao) GetFollower(ctx context.Context, userID int64, pageSize int, pageNum int) ([]*model.User, int, error) {
 	var err error
 
@@ -140,4 +142,32 @@ func (f *FollowerDao) GetFriends(ctx context.Context, userID int64, pageSize int
 	}
 
 	return users, len(followerIDs), nil
+}
+
+func (f *FollowerDao) IsExistFollow(ctx context.Context, userID int64, followerID int64) (bool, error) {
+	count, err := f.q.Follower.WithContext(ctx).
+		Select(f.q.Follower.ID).
+		Where(f.q.Follower.UserID.Eq(userID), f.q.Follower.FollowerID.Eq(followerID)).
+		Limit(1).
+		Count()
+
+	if err != nil {
+		return false, errors.Wrapf(err, "IsExistFollow failed, count: %d, userID: %d", count, userID)
+	}
+	return count > 0, nil
+}
+
+func (f *FollowerDao) IsExistFriend(ctx context.Context, userID int64, friendID int64) (bool, error) {
+	count, err := f.q.Follower.WithContext(ctx).
+		Where(
+			f.q.Follower.UserID.Eq(userID), f.q.Follower.FollowerID.Eq(friendID),
+		).Or(
+		f.q.Follower.UserID.Eq(friendID), f.q.Follower.FollowerID.Eq(userID),
+	).Count()
+
+	if err != nil {
+		return false, errors.Wrapf(err, "IsExistFriend failed, count: %d, userID: %d", count, userID)
+	}
+
+	return count == mutualFollowCount, nil
 }
